Begin proof-of-concept
This is quick, dirty, and gross. The idea is relatively simple - create a Board
object, which contains an array with 32 elements in it. The rest is all game
logic for determining where a given piece can move and whether it can jump
anything.

Apologies for the hilarious couple of 3functional5you functions in there.
I'll refactor them into iterative functions if need be.

Signed-off-by: Michael Bottini <mbottini1@gmail.com>