Add move logic
The game logic takes a sourceIndex and a targetIndex. If all is
successful, it returns a new Board with the new state of the pieces
and a new GameState. If not, it returns undefined.

There is significant redundancy inside this gigantic function, and
a lot of it can be refactored. I just wanted to get it working,
man.