Add check for a piece at targetIndex during move
When we do a regular diagonal move, we have to check to ensure that
the square at targetIndex is not already occupied by a piece.
Currently, moving a piece to an occupied square will overwrite the
piece that is currently there.

Note that this doesn't apply to captures; that logic is already
implemented because we need to do that logic to see if it's
_possible_ to capture.