Change undefined return vals to null
Per Issue #8, `undefined` is the incorrect value to return when you
want to return the Javascript version of `Nothing` from the `Maybe`
monad. We want `null`.

This was *close* to a `%s/undefined/null/g` job. There was one case
(the constructor for the `Board`) where `undefined` was the correct
value, and that was left alone. Everything else is `null`, though.