Add Websocket Game handling
This addition implements the actual work of creating a game room
and providing an interface to create and join games. Creating or
joining a game subscribes the websocket to that game's interface
instead of the game room's interface. When the game ends, the
websocket is resubscribed to the game room's interface.

I'm not sure if anything works yet. Testing must start from here.