Make AGC message parsing more robust
The ActiveGameController currently crashes on any malformed messages
that do not parse to valid JSON or don't have the correct properties.

I have extended the JSONRequest class to include three new classes.
The processMessage function now rejects all malformed requests,
and prints a distinct error message if it got a different JSONRequest
(as if for a Game object).