Add stalemate logic
When a player cannot make any legal moves, they lose. This
iterates through the Board of a particular color and determines
whether any of that color's pieces can move. If it can't,
it ends the game.

Because the logic is now slightly different for determining a
gameOver, the ActiveGame needs slightly different logic for
implementing finishGame.